+++ assets/javascripts/kitten/components/notifications/badge.test.js
@@ -1,0 +1,42 @@
+import React from 'react'
+import { expect } from 'chai'
+import { shallow } from 'enzyme'
+import { Badge } from 'kitten/components/notifications/badge'
+
+describe('<Badge />', () => {
+  describe('by default', () => {
+    const badge = shallow(<Badge />)
+
+    it('is a <span />', () => {
+      expect(badge).to.have.tagName('span')
+    })
+
+    it('has a default classes', () => {
+      expect(badge).to.have.className('k-Badge')
+    })
+  })
+
+  describe('with custom class', () => {
+    const badge = shallow(<Badge className="custom__class" />)
+
+    it('has a custom class', () => {
+      expect(badge).to.have.className('custom__class')
+    })
+  })
+
+  describe('with other prop', () => {
+    const badge = shallow(<Badge aria-hidden="true" />)
+
+    it('has an aria-hidden attribute', () => {
+      expect(badge).to.have.attr('aria-hidden', 'true')
+    })
+  })
+
+  describe('with children', () => {
+    const badge = shallow(<Badge>42</Badge>)
+
+    it('has text', () => {
+      expect(badge).to.have.text('42')
+    })
+  })
+})